Abstracted build comparisons into a parent/child relationship
This was done in anticipation of Accuracy Graphs needing the same setup of 4 builds and a self-contained object that ActionListens to its own JPanel. Ideally this would also let me create other ways to compare weapons later too.
